requirements.txt
altgraph==0.17.3
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
future==0.18.2
google-api-core==2.10.1
google-auth==2.12.0
google-cloud-vision==3.1.3
googleapis-common-protos==1.56.4
grpcio==1.49.1
grpcio-status==1.49.1
idna==3.4
numpy==1.23.3
opencv-python==4.6.0.66
pefile==2022.5.30
Pillow==9.2.0
proto-plus==1.22.1
protobuf==4.21.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyinstaller==5.4.1
pyinstaller-hooks-contrib==2022.10
pywin32-ctypes==0.2.0
requests==2.28.1
rsa==4.9
six==1.16.0
urllib3==1.26.12
